requirements.txt
torch
prettytable
pandas
tqdm
numpy
matplotlib
tensorboardX
python-dotenv
scikit-learn
ccxt
rich
pyyaml
databases
sqlalchemy
asyncpg
psycopg2-binary